requirements.txt
# Core dependencies
PyQt5>=5.15.0
PyQt5-sip>=12.8.0
PyQtChart>=5.15.0
numpy>=1.19.0
opencv-python>=4.5.0
PyYAML>=5.4.0
requests>=2.25.0
psutil>=5.8.0

# Video processing
av>=8.0.0  # PyAV for RTSP

# Object detection
ultralytics>=8.0.0  # YOLOv8
torch>=1.9.0
torchvision>=0.10.0

# Optional dependencies
GPUtil>=1.4.0  # GPU monitoring
pygame>=2.0.0  # Sound alerts

# Development and testing
pytest>=6.2.0
pylint>=2.8.0
black>=21.5b0